+++ docs/whats_new.rst
@@ -30,6 +30,13 @@ Enhancements
   Note that only xarray arrays can be detected as unstructured grids.
   (:issue:`278`, :pull:`280`). By `Aaron Spring <https://github.com/aaronspring>`_.
 
+- The plotting methods (:py:meth:`Regions.plot` and :py:meth:`Regions.plot_regions`) now
+  use a more sophisticated logic to subsample lines on GeoAxes plots. The new method is
+  based on the eucledian distance of each segment. Per default the maximum distance of
+  each segment is 1 for lat/ lon coords - see the ``tolerance`` keyword of the plotting
+  methods. The ``subsample`` keyword is deprecated (:issue:`109`, :pull:`292`).
+
+
 Deprecations
 ~~~~~~~~~~~~